requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
pypdf
openai
tiktoken
streamlit
python-dotenv
